+++ lib/openmaize/login/base.ex
@@ -47,15 +47,17 @@ defmodule Openmaize.Login.Base do
         check_user_pass conn, unpack_params(params), opts
       end
 
+      def unpack_params(%{"email" => email, "password" => password}), do: {:email, email, password}
       def unpack_params(%{"username" => username, "password" => password}) do
         {:username, username, password}
       end
-      def unpack_params(%{"email" => email, "password" => password}), do: {:email, email, password}
+      def unpack_params(_), do: nil
 
       defoverridable [init: 1, call: 2, unpack_params: 1]
     end
   end
 
+  require Logger
   import Plug.Conn
   alias Openmaize.Config
 
@@ -65,27 +67,29 @@ defmodule Openmaize.Login.Base do
   def check_user_pass(conn, {uniq, user_id, password}, {repo, user_model}) do
     repo.get_by(user_model, [{uniq, user_id}])
     |> check_pass(password, Config.hash_name)
-    |> handle_auth(conn)
+    |> handle_auth(conn, user_id)
   end
+  def check_user_pass(_, _, _), do: raise ArgumentError, "invalid params or options"
 
-  defp check_pass(nil, _, _), do: Config.crypto_mod.dummy_checkpw
-  defp check_pass(%{confirmed_at: nil}, _, _),
-    do: {:error, "You have to confirm your email address before continuing."}
+  defp check_pass(nil, _, _) do
+    Config.crypto_mod.dummy_checkpw
+    {:error, "invalid user-identifier"}
+  end
+  defp check_pass(%{confirmed_at: nil}, _, _), do: {:error, "account unconfirmed"}
   defp check_pass(user, password, hash_name) do
     %{^hash_name => hash} = user
-    Config.crypto_mod.checkpw(password, hash) and {:ok, user}
+    Config.crypto_mod.checkpw(password, hash) and
+    {:ok, user} || {:error, "invalid password"}
   end
 
-  defp handle_auth({:ok, %{id: id, otp_required: true}}, conn) do
+  defp handle_auth({:ok, %{id: id, otp_required: true}}, conn, _) do
     put_private(conn, :openmaize_otpdata, id)
   end
-  defp handle_auth({:ok, user}, conn) do
+  defp handle_auth({:ok, user}, conn, _) do
     put_private(conn, :openmaize_user, user)
   end
-  defp handle_auth({:error, message}, conn) do
-    put_private(conn, :openmaize_error, message)
-  end
-  defp handle_auth(_, conn) do
+  defp handle_auth({:error, message}, conn, user_id) do
+    Logger.warn "#{conn.request_path} #{user_id} - #{message}"
     put_private(conn, :openmaize_error, "Invalid credentials")
   end
 end