+++ test/custom_auth_test.exs
@@ -1,6 +1,0 @@
-defmodule Openmaize.CustomAuthTest do
-  use ExUnit.Case
-  use Plug.Test
-
-
-end